[AJA-425] Chore: Mapper에 파라미터 추가

+++ src/main/java/com/newbarams/ajaja/module/plan/dto/PlanResponse.java
@@ -12,7 +12,7 @@ public class PlanResponse {
 		String title,
 		String description,
 		boolean isPublic,
-		int ajajas,
+		long ajajas,
 		List<String> tags,
 		Instant createdAt
 	) {
@@ -23,7 +23,7 @@ public class PlanResponse {
 		Long userId,
 		String nickname,
 		String title,
-		int ajajas,
+		long ajajas,
 		List<String> tags,
 		Instant createdAt
 	) {
